apps/workers/attendance_automation_server/requirements.txt
insightface>=0.7.3
faiss-cpu>=1.7.4
opencv-python>=4.8.0
numpy>=1.24.0
onnxruntime>=1.15.0
